fix(main): drop redundant u32-to-u32 cast in clean_jobserver
GHAS/clippy flagged casting jobserver_tokens(ctx)'s already-u32 result
through `as u32` before widening to u64 — a no-op cast left over from
before jobserver_tokens returned u32 directly.

+++ src/main.rs
@@ -4785,7 +4785,7 @@ pub unsafe fn clean_jobserver(ctx: &crate::execctx::ExecContext, status: i32) {
                 INTSTR_LENGTH,
                 b"INTERNAL: exiting with %u jobserver tokens (should be 0)!\0" as *const u8
                     as *const ::core::ffi::c_char,
-                &[FmtArg::Uint((jobserver_tokens(ctx)) as u32 as u64)],
+                &[FmtArg::Uint(jobserver_tokens(ctx) as u64)],
             );
         } else {
             loop {
